refactor: mv ModuleLog to module_log.go
